fix #26
Patch proposal to fix #26
When DOM elements are detached from the document, offsetWidth/Height are
always 0: whereas the element may have proper size set in width/height
attributes. This patch adds a fallback on width/height properties when
offsetWidth/Height are 0. It allows to use a canvas as a texture without
having to attach it first.