Make the setSize method of RenderingPanel protected
The offsetWidth and offsetHeight may not work as expected in some scenarios. It is important to give a choice to the developer in those cases. We have hit that bug on chrome. It the offsetWidth stopped working when the root panel is in an iframe. Therefore, it could be good to make the setSize method of RenderingPanel  overridable.